Break apart the Data.Save module
Pull out migrations, validations, and save loading
