+++ test/data/save_test.exs
@@ -1,10 +1,12 @@
 defmodule Data.SaveTest do
   use ExUnit.Case
   import TestHelpers
-  doctest Data.Save
 
   alias Data.Save
-  alias Data.Save.Config
+  alias Data.Save.Validations
+
+  doctest Data.Save
+  doctest Data.Save.Validations
 
   describe "loading" do
     test "converts strings to atoms as keys" do
@@ -53,42 +55,5 @@ defmodule Data.SaveTest do
   test "ensures channels is always an array when loading" do
     {:ok, save} = Save.load(%{})
     assert save.channels == []
-  end
-
-  describe "migrate old save data" do
-    test "migrate item_ids to item instances" do
-      save = %{item_ids: [1], version: 1}
-      save = Save.migrate(save)
-
-      assert save.version > 1
-      assert [%{id: 1}] = save.items
-    end
-
-    test "migrates wearing and wielding items" do
-      save = %{wielding: %{right: 1}, wearing: %{chest: 1}}
-      save = Save.migrate(save)
-
-      assert save.version > 2
-      assert %{right: %{id: 1}} = save.wielding
-      assert %{chest: %{id: 1}} = save.wearing
-    end
-
-    test "will migrate as far as it can" do
-      save = %{item_ids: [1]}
-      save = Save.migrate(save)
-
-      assert save.version > 0
-      assert [%{id: 1}] = save.items
-    end
-  end
-
-  describe "migrating config" do
-    test "adds prompt" do
-      save = %{config: %{}}
-
-      save = Save.migrate_config(save)
-
-      assert save.config.prompt == Config.default_prompt()
-    end
   end
 end
